fix(newport): drop unnecessary same-type cast flagged by clippy 1.94.0
`atoi()` returns `i32` (util.rs), so `atoi(&reply) as i32` is an
`i32 -> i32` cast that clippy::unnecessary_cast denies under the CI-
pinned toolchain 1.94.0. Local clippy (1.97.0) does not flag it, which
is why the earlier local pass missed it. Remove the cast; `select_motor`
stays `i32` (both `if` arms are `i32`).

Enumerated the whole defect family with the exact CI toolchain
(`cargo +1.94.0 clippy --workspace --all-targets --exclude d435i
--exclude d435i-ioc`): this is the only unnecessary_cast site in the
workspace. Full `-D warnings` gate now green under 1.94.0.

+++ drivers/motor/newport/src/pmnc.rs
@@ -581,7 +581,7 @@ impl AsynMotor for PmncAxis {
         // 1. Which motor is selected on this driver? (8751 has only motor 0.)
         let select_motor = if self.is_8753() {
             let reply = checked!(ctrl.command(&format!("CHL {d}")));
-            atoi(&reply) as i32
+            atoi(&reply)
         } else {
             0
         };
